general cleanup: import name updates; refactoring; slight style adjustments; icon update
