+++ app/lib/utils.ts
@@ -1,19 +1,15 @@
-export function checkTransparency(
-  ctx: CanvasRenderingContext2D | null,
-  width: number,
-  height: number
-): boolean {
+export function isTransparent(ctx: CanvasRenderingContext2D | null, width: number, height: number) {
   if (ctx !== null) {
     const { data } = ctx.getImageData(0, 0, width, height);
 
     for (let i = 0; i < data.length; i += 4) {
       if (data[i] !== 0) {
+        console.log("IN IF STATEMENT");
         return false;
       }
     }
-
+    console.log("TRUE STATEMENT");
     return true;
   }
-
   return false;
 }
